+++ stac_fastapi/extensions/stac_fastapi/extensions/core/free_text/free_text.py
@@ -25,18 +25,18 @@ class FreeTextConformanceClasses(str, Enum):
 
     # https://github.com/stac-api-extensions/freetext-search?tab=readme-ov-file#basic
     SEARCH = "https://api.stacspec.org/v1.0.0-rc.1/item-search#free-text"
-    COLLECTIONS = "https://api.stacspec.org/v1.0.0-rc.1/collection-search#free-text"
     ITEMS = "https://api.stacspec.org/v1.0.0-rc.1/ogcapi-features#free-text"
+    COLLECTIONS = "https://api.stacspec.org/v1.0.0-rc.1/collection-search#free-text"
 
     # https://github.com/stac-api-extensions/freetext-search?tab=readme-ov-file#advanced
     SEARCH_ADVANCED = (
         "https://api.stacspec.org/v1.0.0-rc.1/item-search#advanced-free-text"
     )
-    COLLECTIONS_ADVANCED = (
-        "https://api.stacspec.org/v1.0.0-rc.1/collection-search#advanced-free-text"
-    )
     ITEMS_ADVANCED = (
         "https://api.stacspec.org/v1.0.0-rc.1/ogcapi-features#advanced-free-text"
+    )
+    COLLECTIONS_ADVANCED = (
+        "https://api.stacspec.org/v1.0.0-rc.1/collection-search#advanced-free-text"
     )
 
 
@@ -57,8 +57,6 @@ class FreeTextExtension(ApiExtension):
     conformance_classes: List[str] = attr.ib(
         default=[
             FreeTextConformanceClasses.SEARCH,
-            FreeTextConformanceClasses.COLLECTIONS,
-            FreeTextConformanceClasses.ITEMS,
         ]
     )
     schema_href: Optional[str] = attr.ib(default=None)
@@ -92,8 +90,6 @@ class FreeTextAdvancedExtension(ApiExtension):
     conformance_classes: List[str] = attr.ib(
         default=[
             FreeTextConformanceClasses.SEARCH_ADVANCED,
-            FreeTextConformanceClasses.COLLECTIONS_ADVANCED,
-            FreeTextConformanceClasses.ITEMS_ADVANCED,
         ]
     )
     schema_href: Optional[str] = attr.ib(default=None)