+++ stac_fastapi/extensions/tests/test_filter.py
@@ -6,6 +6,11 @@ from starlette.testclient import TestClient
 from stac_fastapi.api.app import StacApi
 from stac_fastapi.api.models import create_get_request_model, create_post_request_model
 from stac_fastapi.extensions.core import FilterExtension
+from stac_fastapi.extensions.core.filter import (
+    CollectionSearchFilterExtension,
+    ItemCollectionFilterExtension,
+    SearchFilterExtension,
+)
 from stac_fastapi.types.config import ApiSettings
 from stac_fastapi.types.core import BaseCoreClient
 
@@ -31,7 +36,7 @@ class DummyCoreClient(BaseCoreClient):
         raise NotImplementedError
 
 
-@pytest.fixture
+@pytest.fixture(autouse=True)
 def client() -> Iterator[TestClient]:
     settings = ApiSettings()
     extensions = [FilterExtension()]
@@ -46,8 +51,63 @@ def client() -> Iterator[TestClient]:
         yield client
 
 
-def test_search_filter_post_filter_lang_default(client: TestClient):
+@pytest.fixture(autouse=True)
+def client_multit_ext() -> Iterator[TestClient]:
+    settings = ApiSettings()
+    extensions = [
+        SearchFilterExtension(),
+        ItemCollectionFilterExtension(),
+        # Technically `CollectionSearchFilterExtension`
+        # shouldn't be registered to the application but to the collection-search class
+        CollectionSearchFilterExtension(),
+    ]
+
+    api = StacApi(
+        settings=settings,
+        client=DummyCoreClient(),
+        extensions=extensions,
+        search_get_request_model=create_get_request_model([SearchFilterExtension()]),
+        search_post_request_model=create_post_request_model([SearchFilterExtension()]),
+    )
+    with TestClient(api.app) as client:
+        yield client
+
+
+@pytest.mark.parametrize("client_name", ["client", "client_multit_ext"])
+def test_filter_endpoints_conformances(client_name, request):
+    """Make sure conformances classes are set."""
+    client = request.getfixturevalue(client_name)
+
+    response = client.get("/conformance")
+    assert response.is_success, response.json()
+    response_dict = response.json()
+    conf = response_dict["conformsTo"]
+    assert (
+        "http://www.opengis.net/spec/ogcapi-features-3/1.0/conf/features-filter" in conf
+    )
+    assert "https://api.stacspec.org/v1.0.0-rc.2/item-search#filter" in conf
+    assert client.get("/queryables").is_success
+    assert client.get("/collections/collection_id/queryables").is_success
+
+
+def test_filter_conformances_collection_search(client_multit_ext):
+    """Make sure conformances classes are set."""
+    response = client_multit_ext.get("/conformance")
+    assert response.is_success, response.json()
+    response_dict = response.json()
+    conf = response_dict["conformsTo"]
+    assert (
+        "http://www.opengis.net/spec/ogcapi-features-3/1.0/conf/features-filter" in conf
+    )
+    assert "https://api.stacspec.org/v1.0.0-rc.2/item-search#filter" in conf
+    assert "https://api.stacspec.org/v1.0.0-rc.1/collection-search#filter" in conf
+
+
+@pytest.mark.parametrize("client_name", ["client", "client_multit_ext"])
+def test_search_filter_post_filter_lang_default(client_name, request):
     """Test search POST endpoint with filter ext."""
+    client = request.getfixturevalue(client_name)
+
     response = client.post(
         "/search",
         json={
@@ -61,8 +121,11 @@ def test_search_filter_post_filter_lang_default(client: TestClient):
     assert response_dict["filter_lang"] == "cql2-json"
 
 
-def test_search_filter_post_filter_lang_non_default(client: TestClient):
+@pytest.mark.parametrize("client_name", ["client", "client_multit_ext"])
+def test_search_filter_post_filter_lang_non_default(client_name, request):
     """Test search POST endpoint with filter ext."""
+    client = request.getfixturevalue(client_name)
+
     filter_lang_value = "cql-json"
     response = client.post(
         "/search",
@@ -78,8 +141,11 @@ def test_search_filter_post_filter_lang_non_default(client: TestClient):
     assert response_dict["filter_lang"] == filter_lang_value
 
 
-def test_search_filter_get(client: TestClient):
+@pytest.mark.parametrize("client_name", ["client", "client_multit_ext"])
+def test_search_filter_get(client_name, request):
     """Test search GET endpoint with filter ext."""
+    client = request.getfixturevalue(client_name)
+
     response = client.get(
         "/search",
         params={